+++ modules/backend/formwidgets/RecordFinder.php
@@ -12,6 +12,7 @@ use Backend\Classes\FormWidgetBase;
  *        label: User
  *        type: recordfinder
  *        list: ~/plugins/rainlab/user/models/user/columns.yaml
+ *        title: Find Record
  *        prompt: Click the Find button to find a user
  *        nameFrom: name
  *        descriptionFrom: email
@@ -39,6 +40,11 @@ class RecordFinder extends FormWidgetBase
      * @var string Relation column to display for the description
      */
     public $descriptionFrom;
+
+    /**
+     * @var string Text to display for the title of the popup list form
+     */
+    public $title = 'backend::lang.recordfinder.find_record';
 
     /**
      * @var string Prompt to display if no record is selected.
@@ -75,6 +81,7 @@ class RecordFinder extends FormWidgetBase
     public function init()
     {
         $this->fillFromConfig([
+            'title',
             'prompt',
             'keyFrom',
             'nameFrom',
@@ -150,6 +157,7 @@ class RecordFinder extends FormWidgetBase
         $this->vars['descriptionValue'] = $this->getDescriptionValue();
         $this->vars['listWidget'] = $this->listWidget;
         $this->vars['searchWidget'] = $this->searchWidget;
+        $this->vars['title'] = $this->title;
         $this->vars['prompt'] = str_replace('%s', '<i class="icon-th-list"></i>', e(trans($this->prompt)));
     }