+++ modules/system/assets/ui/js/inspector.datainteraction.js
@@ -53,7 +53,8 @@
 
             try {
                 return valuesStr.length === 0 ? {} : $.parseJSON(valuesStr)
-            } catch (err) {
+            }
+            catch (err) {
                 throw new Error('Error parsing Inspector field values. ' + err)
             }
         }
@@ -113,7 +114,7 @@
             self = this
 
         $.oc.stripeLoadIndicator.show()
-        var request = $form.request('onGetInspectorConfiguration', {
+        $form.request('onGetInspectorConfiguration', {
             data: data
         }).done(function inspectorConfigurationRequestDoneClosure(data) {
             self.configurartionRequestDone(data, onComplete, result)
@@ -134,10 +135,12 @@
 
             try {
                return $.parseJSON(configuration)
-            } catch(err) {
+            }
+            catch(err) {
                 throw new Error('Error parsing Inspector configuration. ' + err)
             }
-        } else {
+        }
+        else {
             return configuration
         }
     }