+++ modules/system/assets/ui/js/inspector.surface.js
@@ -376,7 +376,8 @@
         if (property.itemType == 'group' || isGroupedControl) {
             row.setAttribute('data-group-index', this.getGroupManager().getGroupIndex(group))
             row.setAttribute('data-parent-group-index', this.getGroupManager().getGroupIndex(group.parentGroup))
-        } else {
+        }
+        else {
             if (group.parentGroup) {
                 row.setAttribute('data-parent-group-index', this.getGroupManager().getGroupIndex(group))
             }
@@ -446,7 +447,8 @@
 
                     self.expandOrCollapseRows(rows, collapse, duration, noAnimation)
                 }, duration)
-            } else {
+            }
+            else {
                 $.oc.foundation.element.toggleClass(row, 'collapsed', collapse)
                 $.oc.foundation.element.toggleClass(row, 'expanded', !collapse)
 
@@ -583,8 +585,9 @@
 
     Surface.prototype.findPropertyEditor = function(property) {
         for (var i = 0, len = this.editors.length; i < len; i++) {
-            if (this.editors[i].getPropertyName() == property)
+            if (this.editors[i].getPropertyName() == property) {
                 return this.editors[i]
+            }
         }
 
         return null
@@ -592,8 +595,9 @@
 
     Surface.prototype.findExternalParameterEditor = function(property) {
         for (var i = 0, len = this.externalParameterEditors.length; i < len; i++) {
-            if (this.externalParameterEditors[i].getPropertyName() == property)
+            if (this.externalParameterEditors[i].getPropertyName() == property) {
                 return this.externalParameterEditors[i]
+            }
         }
 
         return null