+++ modules/system/console/PluginRefresh.php
@@ -38,6 +38,9 @@ class PluginRefresh extends Command
     {
         $pluginName = $this->argument('name');
         $pluginName = PluginManager::instance()->normalizeIdentifier($pluginName);
+        if (!PluginManager::instance()->exists($pluginName)) {
+            throw new \InvalidArgumentException(sprintf('Plugin "%s" not found.', $pluginName));
+        }
 
         $manager = UpdateManager::instance()->resetNotes();