+++ tests/unit/cms/classes/CodeParserTest.php
@@ -6,14 +6,14 @@ use Cms\Classes\Layout;
 use Cms\Classes\CodeParser;
 use Cms\Classes\Controller;
 
-class CodeParserTest extends TestCase 
+class CodeParserTest extends TestCase
 {
     public static function getProperty($name)
     {
         $class = new ReflectionClass('\Cms\Classes\CodeParser');
         $property = $class->getProperty($name);
         $property->setAccessible(true);
-     
+
         return $property;
     }
 
@@ -251,7 +251,7 @@ class CodeParserTest extends TestCase
         $obj = $parser->source($page, null, $controller);
         $this->assertInstanceOf('\Cms\Classes\PageCode', $obj);
 
-        $referenceFilePath = base_path().'/tests/fixtures/cms/reference/namespaces.php';
+        $referenceFilePath = base_path().'/tests/fixtures/cms/reference/namespaces.php.stub';
         $this->assertFileExists($referenceFilePath);
         $referenceContents = $this->getContents($referenceFilePath);